+++ lib/service_actor/arguments_validator.rb
@@ -3,6 +3,13 @@
 module ServiceActor::ArgumentsValidator
   module_function
 
+  def validate_origin_name(name, origin:)
+    return unless ServiceActor::Result.instance_methods.include?(name.to_sym)
+
+    raise ArgumentError,
+          "#{origin} `#{name}` overrides `ServiceActor::Result` instance method"
+  end
+
   def validate_error_class(value)
     return if value.is_a?(Class) && value <= Exception