+++ spec/examples/set_output_called_display.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class SetOutputCalledDisplay < Actor
-  output :display
-
-  def call
-    self.display = "Foobar"
-  end
-end