+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Fixes:
   are subclasses of `Exception` (#132)
 - Skip `default` check for actor outputs (#135)
 - Ensure provided `fail_on` arguments are subclasses of `Exception` (#136)
+- Ensure `input`, `output` and `alias_input` names do not collide with
+  `ServiceActor::Result` instance methods (#138)
 
 ## v3.7.0